+++ changelog.md
@@ -9,6 +9,7 @@ SchemaCrawler AI release notes.
 - Add lint function to MCP server (and remove it from 'aichat')
 - Add table-sample function to MCP server
 - Show remarks and data types wherever possible
+- Add function for getting database server information
 
 
 <a name="v16.26.3-4"></a>